finish

+++ src/components/Booking/Booking.js
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react';
+import { Link } from 'react-router-dom';
 import useAuth from '../Firebase/useAuth';
 
 const Booking = () => {
@@ -7,32 +8,35 @@ const email=user?.email;
 const [booking,setBooking]=useState([])
 
 useEffect(()=>{
-    fetch(`http://localhost:5000/booking/${email}`)
+    fetch(`https://intense-fortress-98735.herokuapp.com/booking/${email}`)
     .then(res=>res.json())
     .then(data=>setBooking(data))
 
 },[user?.email])
 
 const handleDeleteBooking = (id) => {
-     fetch(`http://localhost:5000/booking/${id}`, {
-      method: "DELETE",
-    
-    })
-      .then((res) => res.json())
-      .then((result) => {
-          
-        if (result.deletedCount) {
-           
-           alert("Are you sure?")
-           const remaining=booking.filter(bk=>bk._id !==id)
-           setBooking(remaining)
-
-        } 
-      });
+  const proceed=window.confirm("Are you sure, You want to Delete?")
+    if(proceed){
+      fetch(`https://intense-fortress-98735.herokuapp.com/booking/${id}`, {
+        method: "DELETE",
+      
+      })
+        .then((res) => res.json())
+        .then((result) => {
+            
+          if (result.deletedCount) {
+             
+             alert("Successfully Deleted")
+             const remaining=booking.filter(bk=>bk._id !==id)
+             setBooking(remaining)
+  
+          } 
+        });
+    }
   };
-  const handlecomfrom=()=>{
-      alert("Are You sure to conform")
-  }
+  // const handlecomfrom=()=>{
+  //     alert("Are You sure to conform")
+  // }
     return (
        <div className="container  pb-5">
             <div className="d-flex align-items-center justify-content-center ">
@@ -53,7 +57,8 @@ const handleDeleteBooking = (id) => {
                  
                    <div className="d-flex justify-content-between p-2" >
                    
-                    <button className="btn btn-danger mb-2" onClick={handlecomfrom}>Conform</button>
+                    {/* <button className="btn btn-danger mb-2" onClick={handlecomfrom}>Conform</button> */}
+                    <Link to={`/manageorder/${bok._id}`}><button className="btn btn-primary">Conform</button></Link>
                     <button className="btn btn-danger mb-2" onClick={()=>handleDeleteBooking(bok._id)}>Cencel</button>
                    
                    </div>
